Src/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "northstar-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.112.0",
    "uvicorn[standard]==0.30.5",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "sqlalchemy==2.0.34",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.2",
    "redis==5.0.7",
    "rq==1.16.2",
    "httpx==0.27.2",
    "requests==2.32.3",
    "qdrant-client==1.9.2",
    "python-multipart==0.0.9",
    "passlib[bcrypt]==1.7.4",
    "pyjwt[crypto]==2.9.0",
    "python-dotenv==1.0.1",
    "loguru==0.7.2",
    "opentelemetry-sdk==1.25.0",
    "opentelemetry-exporter-otlp==1.25.0",
    "sentence-transformers==3.0.1",
    "torch>=2.2.0",
    "rapidfuzz==3.9.7",
    "blake3>=0.4.1",
    "PyPDF2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.5.5",
    "mypy==1.11.1",
    "pyright==1.1.375",
    "types-redis",
    "types-requests",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "worker*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
show_error_codes = true
ignore_missing_imports = true
plugins = []
